refactor(tests): split analytics.test.ts into 4 focused files
Splits the 2103-line monolith into:
- analytics-basics: analytics module basics, opt-out, getUtmParams, emitClusterInventory, emitAgentProvidersDetected, opt-out persistence, additional emit functions (smoke tests)
- analytics-emit-core: setAnalyticsUserId, initAnalytics, startGlobalErrorTracking, captureUtmParams, updateAnalyticsIds/setAnalyticsUserProperties/markErrorReported edge cases, emitError, emitConversionStep, emitSessionContext, emitDeveloperSession, emitSnoozed
- analytics-emit-extended: send() gating, emitScreenshot*, initAnalytics automated env detection, setAnalyticsOptOut, emitDeveloperSession conditional, setAnalyticsUserId hashing, prediction/data/fixer/error truncation events, emitClusterInventory, emitAgentProvidersDetected bitmask, widget/dashboard/search/install events
- analytics-error-tracking: startGlobalErrorTracking listeners, markErrorReported/dedup integration, captureUtmParams, emitAIPredictions/ConsensusModeToggled, emitPageView, emitCardAdded, modal/drillDown/lens/headlamp/whiteLabel events

Fixes #8624
